Add files via upload
I will comment the code tomorrow. 
can be run and gives referring doctor option to each of the necessary options with room to add more if need be. // results portal 
Havent implemented a back end yet. will use 100% database to interact with backend so will work on database utilization soon!